% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataExtract.R
\name{assembleCohort}
\alias{assembleCohort}
\title{Create a training dataset for a single cohort}
\usage{
assembleCohort(cohortYear, grade, conn, type = "train")
}
\arguments{
\item{cohortYear}{a character string of the format YYYY-YYYY}

\item{grade}{a character representing the grade level in the format 10 or 0X}

\item{conn}{a connection object with the data connections necessary to access
the database}
}
\value{
a data.frame with all the data for a single cohort organized with
each row representing a single student, and each column a measure for that
student in the grade-school-year selected
}
\description{
Create a training dataset for a single cohort
}

