% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataTransform.R
\name{buildTrainingPool}
\alias{buildTrainingPool}
\title{Build a pool of training observations for a grade level}
\usage{
buildTrainingPool(grade, conn)
}
\arguments{
\item{grade}{a numeric representing the grade level we are interested in
finding training cohorts for}

\item{conn}{a connection string which contains links to databases necessary
to extract the data}
}
\value{
a data.frame with all colums generated by assembleCohort and the
number of rows equal to the number of students per cohort X the number of
cohorts available for a given grade
}
\description{
Build a pool of training observations for a grade level
}
\details{
This function identifies the number of grade N cohorts which have
valid graduation outcomes and then pulls all complete observations from those
cohorts using the assembleCohort function. It then stacks all of these cohorts
into a single data.frame suitable for transformation and analysis later. At
the end it calls cleanTrainingPool to clean up the data and enforce business
rules about variable encoding.
}

