% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lds_connect.R
\name{checkCreds}
\alias{checkCreds}
\title{Check credentials and open a channel. \emph{You probably don't need to call
this function directly, it is called by the \code{pull*} functions.}}
\usage{
checkCreds(dsn = NULL)
}
\arguments{
\item{dsn}{The name of the ODBC database containing the data of interest.
  (Check the 'System DSN' tab in the ODBC Data Source Administrator, for
  these)}
}
\value{
a channel object
}
\description{
\code{checkCreds} is an enhanced wrapper to open a channel using RODBC's
\link[RODBC]{odbcConnect} using your credentials that are stored in
\code{"~/.dpi_datacreds"} (see \link{makeCreds} for information on creating
this file for yourself)
}

