% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataTransform.R
\name{cleanDV.DEWSList}
\alias{cleanDV.DEWSList}
\title{Clean up DV in a DEWSList for proper modeling}
\usage{
cleanDV.DEWSList(DEWSList, DV)
}
\arguments{
\item{DEWSList}{a DEWSlist created by makeDEWSList}

\item{DV}{a character representing the name of the dependent variable}
}
\value{
a DEWSList
}
\description{
Clean up DV in a DEWSList for proper modeling
}
\details{
Reassigns the DEWSList[[i]]$class object to be a factor with the
larger group, graduates, as the first level of a factor, and non-graduates
as the second level
}
\note{
Currently only works if DV == "ontime_grad"
}

