% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataTransform.R
\name{cleanTrainingPool}
\alias{cleanTrainingPool}
\title{Cleans the pooled training cases}
\usage{
cleanTrainingPool(pool)
}
\arguments{
\item{pool}{a data.frame}
}
\value{
A data.frame same number of rows as pool, with data recoded. This
includes all of the recoding business rules and should be modified to match
the specific problem at hand.
}
\description{
Cleans the pooled training cases
}

