% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelsearchCaret.R
\name{cleanUp}
\alias{cleanUp}
\title{Clean up the workspace after fitting models}
\usage{
cleanUp()
}
\value{
nothing
}
\description{
a helper function to clean up the workspace after running a modelSearch
}

