% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constructModels.R
\name{constructModels}
\alias{constructModels}
\title{Construct ensemble models from a data.frame of optional models}
\usage{
constructModels(modelResultsName, grade, mode, DV = "grad_ind", nmodels = 8,
  conn)
}
\arguments{
\item{modelResultsName}{a prefix that includes the path and filename of the
file on the disk where the model search results object is stored}

\item{grade}{a numeric representing the grade level of the object}

\item{mode}{a character representing whether the model was run in dev or
in production mode}

\item{DV}{a character representing the name of the dependent variable in the
model}

\item{nmodels}{the number of models to include in an ensemble}

\item{conn}{a connection string pointing to the database where the gatherData
function can extract data from for model building}
}
\value{
Nothing. Writes the model objects to disk
}
\description{
Construct ensemble models from a data.frame of optional models
}
\details{
By default the function builds four kinds of models. It builds ensembles
based on the top N performing models, and on the N most dissimilar models. To
build these it uses both the caretStack and caretEnsemble method.
}

