% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelsearchCaret.R
\name{DEWS_search}
\alias{DEWS_search}
\title{Evaluate and store model performance across a list of models}
\usage{
DEWS_search(DEWSList, DEWScontrol)
}
\arguments{
\item{DEWSList}{a DEWSList created using the makeDEWSList function}

\item{DEWScontrol}{a control object specifying what kind of search to make}
}
\value{
A data.frame with model performance parameters for each of the methods
tested.
}
\description{
This is the workhorse of DEWS as it conducts a search of models
using DEWScontrol on the data in the DEWSList.
}

