% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkgs.R
\name{dewsBootstrap}
\alias{dewsBootstrap}
\title{Function to prepare the environment for model training and model building}
\usage{
dewsBootstrap()
}
\value{
nothing
}
\description{
Function loads all required packages in the correct order and
installs all packages that need to be installed. This minimizes namespace conflicts.
}
\details{
First the package reads all the packages necessary out of the list of
models ready to be trained stored in inst/data/ewsModels.csv. If the model list
changes and this function is re-run, it will now install all necessary packages.
}

