% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelsearchCaret.R
\name{DEWScontrol}
\alias{DEWScontrol}
\title{Set up parameters to control the modelSearch and modelBuild tasks in DEWS}
\usage{
DEWScontrol(mode, grade = NULL)
}
\arguments{
\item{mode}{character vector with three choices - "DEV", "PROD", and "TEST"}

\item{grade}{a numeric specifying the grade level}
}
\value{
A list with elements
\itemize{
\item upsample - a character representing the upsampling technique to use, currently
only SMOTE is supported
\item length - a numeric representing the length of the tuneGrid to be passed to
the tuneLength option in \code{\link{caret}}
\item metric - a character representing the metric to use to identify the
best model, currently ROC is all that is supported
\item core - a numeric specifying the number of cores to use. In DEV, this is
capped at 1/2 the maximum number of available cores, in other modes it is
2/3
\item folds - a numeric representing the number of folds to use in training models
\item reps - numeric representing the number of repeats to use in training models
\item cvmethod - character representing the method to use for cross-validation,
PROD is repeatedcv, dev and test are cv
\item dewsTrControl - a \code{\link{trainControl}} object derived from values
\item mode - a character storing the mode
}
Any of these values can be modified posthoc by editing the list elements
directly.
}
\description{
This function defines three possible streams for the DEWS model
searching and model training routines to take. PROD is the most rigorous, while
DEV is the quickest for debugging and testing purposes. TEST is somewhere in the
middle. Here the
}
\note{
This function is extremely proscriptive and should be modified
based on the user environment.
}

