% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataExtract.R
\name{expectedGrad}
\alias{expectedGrad}
\title{Calculate expected graduation year}
\usage{
expectedGrad(baseyear, grade)
}
\arguments{
\item{baseyear}{a character representing the school year in a given grade for
a given cohort of students in the form 'YYYY-YY' or 'YYYY-YYYY'}

\item{grade}{a character or numeric representing the grade level}
}
\value{
a school year in the format 'YYYY-YY' representing the expected graduation
year for the cohort
}
\description{
Calculate expected graduation year
}
\details{
This expected graduation is purely based on the number of school years
necessary under normal progression to reach grade 12 and the school year that
grade 12 will coincide with.
}

