% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exportPrep.R
\name{exportPreds}
\alias{exportPreds}
\title{A function to export the results of model predictions to a specific format}
\usage{
exportPreds(stub = "EnsembleModels_APRIL", grade = 5, year = "2013-2014",
  conn, prelim = FALSE, cache = TRUE)
}
\arguments{
\item{stub}{path to filename where the predictions in raw form are stored}

\item{grade}{a numeric for the grade of the prediction filename}

\item{year}{a character representing the year in the filename of predictions}

\item{conn}{a connection object to reach the database}

\item{prelim}{a logical indicating if this is the final or preliminary DEWS
run for the year}

\item{cache}{a logical indicating if results should be stored to disk}
}
\value{
a data.frame with key data elements, rescaled to be interpretable,
as well as the final predictions
}
\description{
A function to export the results of model predictions to a specific format
}

