% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predictCaret.R
\name{findBestThresh}
\alias{findBestThresh}
\title{Find the best threshold for a model}
\usage{
findBestThresh(mod, dataList)
}
\arguments{
\item{mod}{}

\item{dataList}{}
}
\value{
A numeric representing the optimal probability threshold cutoff.
}
\description{
Blah blah blah.
}

