% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataTransform.R
\name{findBinary}
\alias{findBinary}
\title{Find binary variables}
\usage{
findBinary(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
a character vector of variable names corresponding to column names in
data which have fewer than three unique values
}
\description{
Find binary variables in a data.frame and return their names.
}
\note{
This function will return constant variables with only one value as
binary as well
}

