% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataTransform.R
\name{findCohortYears}
\alias{findCohortYears}
\title{Find the years available for training based on grade level}
\usage{
findCohortYears(grade)
}
\arguments{
\item{grade}{a numeric representing the grade level we are interested in
finding training cohorts for}
}
\value{
a vector of school years formatted YYYY-YYYY which represent the
grade N cohorts for which graduation data is currently available. This is
done by simple arithmetic and needs to be updated each time the data advances,
e.g. annuallly
}
\description{
Find the years available for training based on grade level
}

