% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataTransform.R
\name{findKeys}
\alias{findKeys}
\title{Find database key variables}
\usage{
findKeys(names)
}
\arguments{
\item{names}{a vector of variable names}
}
\value{
all variable names that have KEY in the name as a character vector
}
\description{
Find database key variables
}

