% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataTransform.R
\name{gatherData.pred}
\alias{gatherData.pred}
\title{Convert a dataset without a dependent variable}
\usage{
gatherData.pred(grade, conn, year, na.omit = TRUE, ...)
}
\arguments{
\item{grade}{a numeric representing the grade level to extract data for}

\item{conn}{a connection object suitable for extracting data}

\item{year}{a character representing the school year to predict for in YYYY-YYYY format}

\item{na.omit}{logical, should rows with an NA be omitted, default is TRUE}
}
\description{
Convert a dataset without a dependent variable
}
\details{
First the function extracts data from buildTrainingPool, then it builds
a model matrix suitable for passing to a transformation and then prediction step.
The key difference between this and the original gatherData is that there are no
valide dependent variable observations for data we want to make predictions on.
This function is modified to account for that.
}
\note{
na.omit must be true for now as no imputation routines have been integrated
}

