% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataTransform.R
\name{gatherData}
\alias{gatherData}
\title{Convert dataset to a model matrix suitable for training in caret}
\usage{
gatherData(grade, conn, DV = "grad_ind", na.omit = TRUE)
}
\arguments{
\item{grade}{a numeric representing the grade level to extract data for}

\item{conn}{a connection object suitable for extracting data}

\item{DV}{a character representing the variable name of the dependent variable}

\item{na.omit}{logical, should rows with an NA be omitted, default is TRUE}
}
\description{
Convert dataset to a model matrix suitable for training in caret
}
\details{
First the function extracts data from buildTrainingPool, then it builds
a model matrix out of the formula DV ~ ., first excluding all "KEY" variables
which represent year, school, district, and student identifiers. Then it returns
the model matrix.
}
\note{
Using na.omit=FALSE may result in a non-conformable array due to a mismatch
between the DV vector length and data vector length. This function has not been
modified to allow this yet.
}

