% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelsearchCaret.R
\name{getCandidateMethods}
\alias{getCandidateMethods}
\title{Build a list of methods to evaluate performance of}
\usage{
getCandidateMethods(n = NULL, full = FALSE)
}
\arguments{
\item{n}{numeric, number of models to extract}

\item{full}{logical, get all methods or just a small subset}
}
\value{
vector of character strings that represent methods in caret
}
\description{
Identify the caret methods for training DEWS models on. This can
build a list out of the data built into EWStools or restrict to a small subset
of models.
}
\note{
Excludes all methods that require Java by default. Restricted list
includes the list used in the Wisconsin implementation of DEWS as of June 2015.
}

