% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkgs.R
\name{installPackages}
\alias{installPackages}
\title{Install packages in a list, but do not load them}
\usage{
installPackages(package, load = FALSE)
}
\arguments{
\item{package}{character, a package name}

\item{load}{logical, should the package be loaded after installing? default is FALSE}
}
\value{
nothing
}
\description{
Install packages in a list, but do not load them
}

