% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lds_connect.R
\name{lds_connect}
\alias{lds_connect}
\title{Connect to Wisconsin DPI data systems}
\usage{
lds_connect()
}
\value{
a list with four connection channels
}
\description{
Connect to Wisconsin DPI data systems
}
\details{
channel = ldsdb, channel2 = prd1, channel3 = prd2, channel4=DWRSQL_DEV
}

