% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lds_connect.R
\name{makeCreds}
\alias{makeCreds}
\title{Makes a properly formatted credential file}
\usage{
makeCreds(dsn, login, password, replace = FALSE)
}
\arguments{
\item{dsn}{The name (or vector of names) of the ODBC database containing the
  data of interest. If your login and password are identical for all
  databases, use 'ALL' (Check the 'System DSN' tab in the ODBC Data Source
  Administrator, for these).}

\item{login}{The login (or vector of logins) to the dsn.}

\item{password}{The password (or vector of passwords) to the dsn.}

\item{replace}{A logical scalar specifying whether or not \code{makeCreds}
  should overwrite an existing credential file. If a credential file exists
  and \code{replace}=FALSE then \code{makeCreds} exits with an error.}
}
\value{
Nothing. A message is printed if writing the file was successfull.
}
\description{
\code{makeCreds} creates the file used by the \code{pull*} functions to
supply the SQL database with your credentials. Credentials for multiple
databases can be submitted as vectors for each of \code{dsn}, \code{login},
and \code{password}. When submitting multiple credentials make sure that
the order is correct.
}

