% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataTransform.R
\name{makeDEWSList}
\alias{makeDEWSList}
\title{Construct a DEWSlist from a model matrix}
\usage{
makeDEWSList(matrix, DV = NULL, preProcess = TRUE, MODE = "PROD",
  PRELIM = FALSE, P, PVALID)
}
\arguments{
\item{matrix}{a matrix produced by gatherData}

\item{DV}{a character representing the column name associated with the dependent
variable}

\item{preProcess}{logical, should a preProcess routine be applied to the data}

\item{P}{numeric, 0 to 1, proportion of data to be used for training}

\item{PVALID,}{0 to 1, proportion of the data to be used for the validation data}

\item{mode}{character, is this running in production or development mode, see details}

\item{prelim}{logical, is this the preliminary or final model, see details}
}
\value{
a list of lists, the lists are named "traindata", "testdata" and
"validdata", and each list has two elements, a predictor matrix called "preds"
and an outcome vector called "class". A fourth element called "scale" is
included if preProcess = TRUE which contains the preProcess object used
}
\description{
Construct a DEWSlist from a model matrix
}
\details{
This function makes the data suitable for use in EWStools functions
for testing and selecting models.
}

