% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelsearchCaret.R
\name{modelSearch}
\alias{modelSearch}
\title{A wrapper function to simplify the DEWS search process}
\usage{
modelSearch(grade, mode = "PROD", fileName, DV = "grad_ind")
}
\arguments{
\item{grade}{numeric corresponding to the cohort grade}

\item{mode}{a single character vector, one of "DEV", "PROD", or "TEST" for DEWSControl}

\item{fileName}{a character for a stub of a filename to store results to}

\item{DV}{a character with the dependent variable name}
}
\value{
Nothing, stores the results of a DEWS_search to disk
}
\description{
A wrapper function to simplify the DEWS search process
}

