% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constructModels.R
\name{modelSearchResults}
\alias{modelSearchResults}
\title{Parse results of a model search}
\usage{
modelSearchResults(modelResultsName, grade, mode)
}
\arguments{
\item{modelResultsName}{a prefix that includes the path and filename of the
file on the disk where the model search results object is stored}

\item{grade}{a numeric representing the grade level of the object}

\item{mode}{a character representing whether the model was run in dev or
in production mode}
}
\value{
Returns a data.frame with the best fitting models.
}
\description{
Parse results of a model search
}

