% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataExtract.R
\name{peerStats}
\alias{peerStats}
\title{Calculate school level peer attributes for a cohort dataset}
\usage{
peerStats(cohortDF)
}
\arguments{
\item{cohortDF}{a dataframe produced by the pullCohort function}
}
\value{
a data.frame with nrows = number of schools
}
\description{
Calculate school level peer attributes for a cohort dataset
}
\details{
This function calculates an array of peer level demographics for a
cohort of students in a given grade-year. These include means and standard
deviations of test scores, suspension and expulsion rates, attendance rates,
and percent non-white, percent SwD, and percent economically disadvantaged
}

