% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exportPrep.R
\name{postProcess.pred}
\alias{postProcess.pred}
\title{Postprocessing of DEWS predictions}
\usage{
postProcess.pred(predCohort, preProcess)
}
\arguments{
\item{predCohort}{a data.frame produced by the prediction function}

\item{preProcess}{a preProcess object used to preProcess predictions}
}
\value{
a data.frame with variables rescaled to be on their original
interpretable scales
}
\description{
Postprocessing of DEWS predictions
}

