% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataTransform.R
\name{preProcess.DEWSList}
\alias{preProcess.DEWSList}
\title{PreProcess DEWS list}
\usage{
\method{preProcess}{DEWSList}(DEWSList)
}
\arguments{
\item{DEWSList}{a list produced by the makeDEWSList function}
}
\value{
a DEWSList with the additional list element "scale" which includes
the preProcess object
}
\description{
PreProcess DEWS list
}
\details{
The preProcess function is applied to all continuous variables in
the training data. A separate scaling is done to binary variables as well.
This training data pre-processing is then applied to test and validdata so that
all data elements have been scaled relative to the training data.
}

