% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataTransform.R
\name{preProcess.pred}
\alias{preProcess.pred}
\title{Preprocess prediction cohort}
\usage{
\method{preProcess}{pred}(predCohort, preProcess, varNames)
}
\arguments{
\item{predCohort}{a data.frame produced by}

\item{preProcess}{a preProcess object with variable names matching those in predCohort}

\item{varNames}{a vector of variable names to preProcess}
}
\value{
a data.frame of of the same dimensions of predCohort that is rescaled
according to the attributes in the preProcess object
}
\description{
Preprocess prediction cohort
}

