% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataExtract.R
\name{pullCohort}
\alias{pullCohort}
\title{Create a cohort set of students with full data}
\usage{
pullCohort(cohortYear, grade, conn, type = "train")
}
\arguments{
\item{cohortYear}{a character string of the format YYYY-YYYY}

\item{grade}{a character representing the grade level in the format 10 or 0X}

\item{conn}{a connection object with the data connections necessary to access
the database}
}
\value{
Returns a data.frame with student mobility data and merge keys
}
\description{
Create a cohort set of students with full data
}
\details{
This function calls pullDemog, pullWSAS, pullDiscipline, pullMobility
and combines the results based on the unique student key to create a single
student-year-grade record which includes all of the information on a student
by row including assessment scores, attendance, demographics, mobility, and
discipline results. Returns incomplete records for students missing data
in any element, but uses the results of pullDemog as the full list of students.
}

