% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataExtract.R
\name{pullDiscipline}
\alias{pullDiscipline}
\title{Pull down discipline data}
\usage{
pullDiscipline(cohortYear, grade, conn)
}
\arguments{
\item{cohortYear}{a character string of the format YYYY-YYYY}

\item{grade}{a character representing the grade level in the format 10 or 0X}

\item{conn}{a connection object with the data connections necessary to access
the database}
}
\value{
Returns a data.frame with student discipline data and merge keys
}
\description{
Pull down discipline data
}
\details{
This function returns the discipline records for a cohort of students
in given grade and year. Students with no discipline record are not returned.
}

