% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataExtract.R
\name{pullGrad}
\alias{pullGrad}
\title{Pull down graduation data}
\usage{
pullGrad(cohortYear, grade, conn)
}
\arguments{
\item{cohortYear}{a character string of the format YYYY-YYYY}

\item{grade}{a character representing the grade level in the format 10 or 0X}

\item{conn}{a connection object with the data connections necessary to access
the database}
}
\value{
Returns a data.frame with student graduation data and merge keys
}
\description{
Pull down graduation data
}
\details{
This function returns the graduation outcome for a cohort of students
in given grade and year. The graduation outcomes are recoded based on DPI
specific exit codes. There is also logic in this function to eliminate
duplicated records and to parse early graduation.
}

