% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataExtract.R
\name{pullMobility}
\alias{pullMobility}
\title{Pull down mobility data}
\usage{
pullMobility(cohortYear, grade, conn)
}
\arguments{
\item{cohortYear}{a character string of the format YYYY-YYYY}

\item{grade}{a character representing the grade level in the format 10 or 0X}

\item{conn}{a connection object with the data connections necessary to access
the database}
}
\value{
Returns a data.frame with student mobility data and merge keys
}
\description{
Pull down mobility data
}
\details{
This function returns the number of schools and number of districts
attended by a student #' in given grade and year for all students.
}

