% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataExtract.R
\name{pullWSAS}
\alias{pullWSAS}
\title{Pull down assessment data}
\usage{
pullWSAS(cohortYear, grade, conn)
}
\arguments{
\item{cohortYear}{a character string of the format YYYY-YYYY}

\item{grade}{a character representing the grade level in the format 10 or 0X}

\item{conn}{a connection object with the data connections necessary to access
the database}
}
\value{
Returns a data.frame with student assessment data and merge keys
}
\description{
Pull down assessment data
}
\details{
This function is meant to pull a single year and single grade of
assessment data for students and to do the necessary data cleaning and reshaping
to create data with a single row per student per year per grade and with columns
representing assessment results on mathematics and reading. This function is
called as part of a larger function which assembles all data on a cohort for
processing in a model search.
}

