% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataExtract.R
\name{pullWSN}
\alias{pullWSN}
\title{Extract a student WSN for crosswalk and matching purposes}
\usage{
pullWSN(conn)
}
\arguments{
\item{conn}{a connection object with the data connections necessary to access
the database}
}
\value{
a data.frame with all student LDS_KEYS and WSNs for merging purposes
}
\description{
Extract a student WSN for crosswalk and matching purposes
}

