% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predictCaret.R
\name{rawPreds}
\alias{rawPreds}
\title{Generate predictions for a new cohort of students using a model trained by DEWS}
\usage{
rawPreds(stub, grade, conn, mode = "PROD", year = "2013-2014",
  findBest = TRUE, cache = FALSE)
}
\arguments{
\item{stub}{}

\item{grade}{}

\item{mode}{}

\item{year}{}

\item{findBest}{}

\item{cache}{}
}
\value{
Nothing.
}
\description{
Blah blah
}

