% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataTransform.R
\name{scaleCenter}
\alias{scaleCenter}
\title{Scale and center a dataframe given a preProcessing object}
\usage{
scaleCenter(data, preProc, binVars, conVars)
}
\arguments{
\item{data}{a data.frame}

\item{preProc}{a preProcessing object}

\item{binVars}{a vector of column indices representing binary variables}

\item{conVars}{a vector of column indices representing continuous variables in
the preProcessing object}
}
\description{
Scale and center a dataframe given a preProcessing object
}
\details{
Continuous variables are processed by preProcess in the caret package,
binary variables are processed using the rescale function in arm which centers
binary inputs
}

