% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataTransform.R
\name{smote}
\alias{smote}
\title{SMOTE the data for class balance?}
\usage{
smote(dewsList)
}
\arguments{
\item{dewsList}{a single element of a DEWSList, e.g. traindata, testdata}
}
\value{
an element of a dewslist with objects preds and class
}
\description{
Applies the synthetic minority over-sampling technique to a
list in a DEWSList to make the data less unbalanced between classes.
}
\details{
See the documentation for SMOTE in the DMwR package for details on
how the SMOTE function works.
}

