% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exportPrep.R
\name{uncenter}
\alias{uncenter}
\title{Uncenter}
\usage{
uncenter(x)
}
\arguments{
\item{x}{a vector with two distinct values that was previously a binary variable}
}
\value{
a vector, length x, with two distinct values, 0 and 1
}
\description{
A function to uncenter binary variables that have been centered
using the arm package
}
\examples{
x <- rbinom(100, 1, prob = .3)
scaleX <- x - mean(x)
identical(x, as.integer(uncenter(scaleX)))
}

