% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exportPrep.R
\name{unPreProc}
\alias{unPreProc}
\title{Reverse pre-processing done by caret to a dataset}
\usage{
unPreProc(preProc, data)
}
\arguments{
\item{preProc}{a preProcess object from the caret package}

\item{data}{a data.frame with the same elements as the preProcess object}
}
\value{
a data.frame
}
\description{
Reverse pre-processing done by caret to a dataset
}
\details{
This only reverses scaling and centering done by preProcess currently.
It cannot undo PCA transformations, it cannot undo imputation, exponential transformations,
or any other method
}

